+++ api/config.py
@@ -1,0 +1,1 @@
+LIVECOINWATCH_API_KEY = "d960befc-10d2-4630-93cf-136f2e7f1558"
